refactor: 必要な部分だけ utils に切り出し
